Refactored method locations.
